chore: move signing stuff to a new folder
